Improve manifest fetching
If the database file is already present, don't fetch it (this will
likely need to change as Destiny gets updates, but for now it's fine)

Use a temporary file for the delivered zip instead of doing a manual
temp file.

+++ get_manifest.py
@@ -1,7 +1,13 @@
 import os
-import pprint
 import requests
+import sys
+import tempfile
 import zipfile
+
+try:
+    import pathlib
+except ImportError:
+    import pathlib2 as pathlib
 
 headers = {'X-API-Key': os.getenv('BUNGIE_API_KEY')}
 
@@ -9,20 +15,25 @@ MANIFEST_URL = 'https://bungie.net/Platform/Destiny2/Manifest/'
 
 response = requests.get(MANIFEST_URL, headers=headers)
 manifest = response.json()
-pprint.pprint(manifest)
 manifest_file_url = 'http://www.bungie.net' + manifest['Response']['mobileWorldContentPaths']['en']
 
+target = pathlib.Path('./Manifest.db')
+
+if target.exists():
+    print("Manifest.db already present, nothing to do!")
+    sys.exit(0)
+
+temp_zip = tempfile.NamedTemporaryFile(mode='wb')
 
 r = requests.get(manifest_file_url)
-with open('MANZIP', 'wb') as zip:
-    zip.write(r.content)
+temp_zip.write(r.content)
 print("Download complete")
 
 
 # Extract the file contents, and rename the extracted file
 # to 'Manifest.content'
-with zipfile.ZipFile('MANZIP') as zip:
+with zipfile.ZipFile(temp_zip.name) as zip:
     name = zip.namelist()
     zip.extractall()
 os.rename(name[0], 'Manifest.db')
-print('Unzipped!')
+print('Unzipped as Manifest.db')
